Remove old pdf export and typst dependency
