+++ README.md
@@ -18,6 +18,4 @@ Rustwell currently implements the entirely of the *Fountain* specification. It o
 
 The original CSS code for the html output was created by [Jonathan Poritsky](https://www.candlerblog.com/), but has been expanded upon for this endeavour.
 
-The typst template is based on the [celluloid](https://typst.app/universe/package/celluloid) template for typst by [Casey Dahlin](mailto:sadmac@google.com), licensed under [Apache-2.0](http://www.apache.org/licenses/LICENSE-2.0) license. The template used for Rustwell is a heavily modified version of celluloid, and is not considered fit for use outside of the scope of Rustwell. The original license can be found [here](./licenses/OFL-Courier-Prime.txt).
-
 [Courier Prime](https://quoteunquoteapps.com/courierprime/), designed by Alan Dague-Greene, is distributed along with Rustwell for the `pdf`-exporter to have a baseline font. The original license can be found [here](./licenses/OFL-Courier-Prime.txt).
