+++ crates/rustwell/src/export.rs
@@ -10,8 +10,6 @@
 
 pub mod html;
 pub mod pdf;
-pub mod pdf2;
-pub mod typst;
 
 use std::{
     fs::File,
